Added pageable protocol
